Added implementation, test data and test cases for fetchFact service call
git-svn-id: https://svn.kuali.org/repos/student/brms/brms-impl/branches/brms-impl-dev@2418 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ fact-finder-impl/src/main/java/org/kuali/student/rules/factfinder/entity/LUIPerson.java
@@ -33,7 +33,7 @@ public class LUIPerson {
     
     private String cluId;
     
-    private Long credits;
+    private Double credits;
 
     
     /**
@@ -89,14 +89,14 @@ public class LUIPerson {
     /**
      * @return the credits
      */
-    public Long getCredits() {
+    public Double getCredits() {
         return credits;
     }
 
     /**
      * @param credits the credits to set
      */
-    public void setCredits(Long cretids) {
+    public void setCredits(Double cretids) {
         this.credits = cretids;
     }    
 }
